+++ EXAMPLES/small_benchmark_run_to_test_very_simple_Earth/run_this_example.sh
@@ -31,22 +31,35 @@ ln -s ../../../DATA/QRFSI12
 ln -s ../../../DATA/topo_bathy
 cd ../
 
-# compiles executables in root directory
-# using default configuration
-cp DATA/Par_file ../../DATA
-cd ../../
-make clean
-make -j4 all
+# checks if executables were compiled and available
+if [ ! -e ../../bin/xspecfem3D ]; then
+  echo "Compiling first all binaries in the root directory..."
+  echo
 
-# checks exit code
-if [[ $? -ne 0 ]]; then exit 1; fi
+  # compiles executables in root directory
+  # using default configuration
+  cd ../../
 
-# backup of constants setup
-cp setup/* $currentdir/OUTPUT_FILES/
-cp OUTPUT_FILES/values_from_mesher.h $currentdir/OUTPUT_FILES/values_from_mesher.h.compilation
-cp DATA/Par_file $currentdir/OUTPUT_FILES/
+  # only in case static compilation would have been set to yes in Makefile:
+  cp $currentdir/DATA/Par_file DATA/Par_file
+  sed -i "s:SAVE_FORWARD.*:SAVE_FORWARD                    = .true.:"  DATA/Par_file
 
-cd $currentdir
+  # compiles code
+  make clean
+  make -j4 all
+
+  # checks exit code
+  if [[ $? -ne 0 ]]; then exit 1; fi
+
+  # backup of constants setup
+  cp setup/* $currentdir/OUTPUT_FILES/
+  if [ -e OUTPUT_FILES/values_from_mesher ]; then
+    cp OUTPUT_FILES/values_from_mesher.h $currentdir/OUTPUT_FILES/values_from_mesher.h.compilation
+  fi
+  cp DATA/Par_file $currentdir/OUTPUT_FILES/
+
+  cd $currentdir
+fi
 
 # copy executables
 mkdir -p bin
